requirements.txt
# Core ML/NLP Libraries
transformers==4.38.2
sentence-transformers==2.2.2
torch==2.1.0
scikit-learn==1.3.2

# Embedding & Retrieval
faiss-cpu==1.7.4

# NLP Models and Pipelines
spacy==3.5.0
scispacy==0.5.5

# Data Handling
pandas==2.2.1
numpy==1.24.4

# Utility
tqdm==4.66.1
gdown==4.7.1
requests==2.31.0
python-dotenv==1.1.1
streamlit==1.36.0
